refactor: split validation module by responsibility
- Split src/validation.ts into focused modules:
  - validators.ts: Basic validation functions (77 lines)
  - factory.ts: ResponsiveFocalPoint creation (40 lines)
  - media-query.ts: Media query generation (20 lines)

- Update import paths to be explicit instead of using re-exports
- Maintain 100% test coverage (74 tests passing)
- Improve code readability and maintainability

BREAKING CHANGE: Import paths for validation functions have changed
